breaking changes: default now does NOT attack
- change in specification
- this guarantees that even caacbr has no cycles
- fixes issues raised by other cases which are equal to default
- fixes a previous bug: case inconsistent with default was not being added

There are two alternatives that go around this, and would have the
same behaviour before or after this breaking change:
Alternative 1) to this would be always replacing the default, but then
behaviour is not specified when there are multiple cases with this
same characterisation/factorset.
Alternative 2) is that default case should be the unique minimum in
the casebase, so no other case has the same characterisation.